requirements.txt
flask>=2.0.0
tensorflow>=2.12.0
Pillow>=9.0.0
numpy>=1.22.0
elasticsearch>=8.5.0